requirements.txt
# Multi-Agent Research Assistant Dependencies

# Core Agent Frameworks
crewai==0.55.2
crewai-tools==0.8.3
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-chroma==0.1.2

# Web Search and Data Retrieval
duckduckgo-search==6.2.3
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.45.0

# Vector Database and RAG
chromadb==0.4.24
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# Memory and State Management
redis==5.0.7
sqlite3-to-mysql==2.1.7

# Safety and Validation
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0

# API and Integration
openai>=1.40.0
anthropic==0.31.2
httpx==0.27.0

# Utilities
python-json-logger==2.0.7
rich==13.7.1
pyyaml==6.0.1

